[REFACTOR] TimelineMapper 리팩터링 #206
